feat: move all logic to service
